Extract interface from PublishQueue, add JavaDoc
PI-1690
