+++ Core/src/com/serotonin/m2m2/rt/publish/mock/MockPublisherRT.java
@@ -4,8 +4,6 @@
 
 package com.serotonin.m2m2.rt.publish.mock;
 
-import com.serotonin.m2m2.rt.dataImage.PointValueTime;
-import com.serotonin.m2m2.rt.publish.PublishQueue;
 import com.serotonin.m2m2.rt.publish.PublisherRT;
 import com.serotonin.m2m2.rt.publish.SendThread;
 import com.serotonin.m2m2.vo.publish.mock.MockPublishedPointVO;
@@ -30,11 +28,6 @@ public class MockPublisherRT extends PublisherRT<MockPublisherVO, MockPublishedP
     public void terminateImpl() {
 
     }
-
-    PublishQueue<MockPublisherVO, MockPublishedPointVO, PointValueTime> getPublishQueue() {
-        return queue;
-    }
-
 
     class MockSendThread extends SendThread {
